MWPW-188979 | Style cleanup - splitting the huge css into operation specific ones
MWPW-188979 | Style cleanup - splitting the huge css into operation specific ones
