+++ streamlibs/operations/preflight/preflight.css
@@ -1,0 +1,58 @@
+/* ============ PREFLIGHT OPERATION =========== */
+  
+.preflight-iframe {
+  width: 100%;
+  height: 100%;
+  border: none;
+}
+
+.show-controls .button-container,
+.show-controls .edit-actions {
+  display: flex;
+}
+
+.preflight-operation-container {
+  font-family: 'Adobe Clean', 'Trebuchet MS', sans-serif;
+  min-height: 100vh;
+  width: 100%;
+  background: #f8f9fa;
+  padding: 40px;
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-direction: column;
+  gap: 20px;
+}
+
+.preflight-operation-container h1 {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 1.5rem;
+  color: var(--sm-color-primary-dark);
+  padding: 20px 32px;
+  background: rgba(255, 255, 255, 0.7);
+  border: 1px solid rgba(59, 130, 246, 0.15);
+  border-radius: 100px;
+  backdrop-filter: blur(20px);
+  box-shadow: 
+      0 4px 24px rgba(59, 130, 246, 0.12),
+      0 1px 2px rgba(0, 0, 0, 0.04),
+      inset 0 1px 0 rgba(255, 255, 255, 0.8);
+  max-width: 90%;
+}
+
+.preflight-operation-actions {
+  font-size: 16px;
+  display: flex;
+  gap: 20px;
+  width: 100%;
+  align-items: center;
+  justify-content: center;
+}
+
+.preflight-operation-actions a {
+  text-decoration: none;
+  color: var(--sm-color-primary); 
+}
