+++ streamlibs/previewer.js
@@ -8,8 +8,6 @@ import {
   pushTargetHtmlToStore,
   fetchPreviewHtmlFromStore,
   pushPreviewHtmlToStore,
-  resetTargetHtmlInStore,
-  resetPreviewHtmlInStore,
 } from './store/store.js';
 import {
   getQueryParam,
@@ -28,8 +26,6 @@ import {
 } from './utils/operations.js';
 import { LOADER_PROGRESS_STEPS, LOADER_STEP_MESSAGES } from './utils/constants.js';
 import { initializeLoader, updateLoader, hideLoader } from './utils/loader.js';
-
-let BUTTON_CONTAINER = null;
 
 function hideDOMElements(eles = []) {
   if (!eles.length) return;
@@ -94,89 +90,6 @@ async function paintHtmlOnPage() {
   mainEle.innerHTML = fetchPreviewHtmlFromStore();
   document.body.prepend(mainEle);
   document.body.prepend(headerEle);
-  if (!BUTTON_CONTAINER) {
-    const div = document.createElement('div');
-    div.classList.add('button-container');
-    const pushToDABtn = createPushButton();
-    const openInDABtn = createOpenButton();
-    const backToEditBtn = createBackToEditButton();
-    const preflightBtn = createPreflightButton();
-    div.append(...[pushToDABtn, openInDABtn, preflightBtn]);
-    document.body.append(div);
-    pushToDABtn.addEventListener('click', handlePushClick);
-    if (backToEditBtn) {
-      div.prepend(backToEditBtn);
-      backToEditBtn.addEventListener('click', handleBackToEditClick);
-    }
-    preflightBtn.addEventListener('click', handlePreflightClick);
-    BUTTON_CONTAINER = div;
-  }
-}
-
-function createPushButton() {
-  const button = document.createElement('a');
-  button.href = '#';
-  button.classList.add('cta-button');
-  button.innerHTML = '<span class="da-push-icon"></span><span class="text">Push to DA</span>';
-  return button;
-}
-
-function createOpenButton() {
-  const button = document.createElement('a');
-  const targetUrl = `https://da.live/edit#${window.streamConfig.targetUrl.startsWith('/') ? window.streamConfig.targetUrl : `/${window.streamConfig.targetUrl}`}`;
-  button.href = targetUrl;
-  button.target = '_blank';
-  button.classList.add('cta-button');
-  button.id = 'open-in-da-button';
-  button.innerHTML = '<span class="da-open-icon"></span><span class="text">Open in DA</span>';
-  if (window.streamConfig.operation === 'create') button.classList.add('disabled');
-  return button;
-}
-
-function createBackToEditButton() {
-  if (window.streamConfig.operation !== 'edit') return;
-  const button = document.createElement('a');
-  button.href = '#';
-  button.classList.add('cta-button');
-  button.id = 'back-to-edit-button';
-  button.innerHTML = '<span class="da-edit-icon"></span><span class="text">Back to Editor</span>';
-  // eslint-disable-next-line consistent-return
-  return button;
-}
-
-function createPreflightButton() {
-  const button = document.createElement('a');
-  button.href = '#';
-  button.classList.add('cta-button');
-  button.innerHTML = '<span class="da-preflight-icon"></span><span class="text">Preview and Preflight</span>';
-  return button;
-}
-
-async function handlePushClick(event) {
-  const button = event.target.closest('.cta-button');
-  const buttonIcon = button.querySelector('span.da-push-icon');
-  buttonIcon.classList.add('sending');
-  try {
-    await persist();
-  } catch (error) {
-    console.log('Error persisting content', error);
-  }
-  buttonIcon.classList.remove('sending');
-  document.querySelector('#open-in-da-button').classList.remove('disabled');
-}
-
-// eslint-disable-next-line no-unused-vars
-async function handleBackToEditClick(event) {
-  BUTTON_CONTAINER.style.display = 'none';
-  resetTargetHtmlInStore();
-  resetPreviewHtmlInStore();
-  document.querySelector('header').remove();
-  document.querySelector('main').remove();
-  await handleBackToEditor();
-}
-
-async function handlePreflightClick() {
-  await preflightOperation();
 }
 
 async function requestStreamConfigFromParent() {
