+++ streamlibs/utils/error-handler.js
@@ -1,24 +1,24 @@
 export function showErrorPage(context = '') {
   const safeContext = (context || 'processing your request').replace(/[<>]/g, '');
   document.body.innerHTML = `
-      <div class="enigma-error-page">
-          <div class="enigma-error-card">
-            <div class="enigma-error-image-wrap">
+      <div class="stream-error-page">
+          <div class="stream-error-card">
+            <div class="stream-error-image-wrap">
               <img
                 src="${window.location.origin}/streamlibs/assets/error-image.webp"
                 alt="Something went wrong"
               >
             </div>
-            <div class="enigma-error-copy">
-                <h1 class="enigma-error-title">Oops! Something broke while ${safeContext}</h1>
-                <p class="enigma-retry-line">
-                  <span class="enigma-retry-text">Give it another go?</span>
-                  <button type="button" id="enigma-retry-btn" class="enigma-retry-btn">Yes Retry</button>
+            <div class="stream-error-copy">
+                <h1 class="stream-error-title">Oops! Something broke while ${safeContext}</h1>
+                <p class="stream-retry-line">
+                  <span class="stream-retry-text">Give it another go?</span>
+                  <button type="button" id="stream-retry-btn" class="stream-retry-btn">Yes Retry</button>
                 </p>
             </div>
           </div>
       </div>`;
-  const retryButton = document.querySelector('#enigma-retry-btn');
+  const retryButton = document.querySelector('#stream-retry-btn');
   retryButton?.addEventListener('click', () => {
     retryButton.disabled = true;
     retryButton.classList.add('is-loading');
