+++ streamlibs/utils/operations.js
@@ -1,10 +1,10 @@
 /* eslint-disable no-use-before-define */
 /* eslint-disable no-console */
 
-export { createStreamOperation } from '../operations/create.js';
+export { createStreamOperation } from '../operations/create/create.js';
 export {
   editStreamOperation,
   applyEditChanges,
   handleBackToEditor,
-} from '../operations/edit.js';
-export { preflightOperation } from '../operations/preflight.js';
+} from '../operations/edit/edit.js';
+export { preflightOperation } from '../operations/preflight/preflight.js';
